refactor: Extract a behaviour for the OAuth2 adapter
So we can add a fake adapter to test the multi-user feature to be implemented.
